Propagate exit code of failed sidekiq startup
Previously, this script would exit with 0 even if sidekiq failed to
start, causing GDK to emit an event with exit_code=0.

This also fixes passing options to sidekiq cluster.

Previously, passing options like `--timeout 10` to `bin/background_jobs`
would result in the following call:

  bin/sidekiq-cluster [...] -e --timeout 10

and not

  bin/sidekiq-cluster [...] -e "" --timeout 10

This made `-e=--timeout` and `10` being ignored.