Display correct code owners when viewing a tree
When viewing a tree blob we currently show the incorrect code owners.

If we view https://gitlab.com/gitlab-org/gitlab/-/tree/master/doc-locale
we will see all of the owners for `*` instead of the owners for
`/doc-locale/` because the git records the path of tree blobs as the
folder name without any path seperators i.e. `doc-locale` instead of
`./doc-locale/`.

The CodeOwners logic needs a path relative to the repository root so we
need to wrap the `blob.path` value in `/` if the blob is a tree.

This change adds:

1. `Blob#tree?` which returns `true` when the `Blob#mode` is `'40000'`.
2. Wraps blob.path when calling Gitlab::CodeOwners::Loader.for_blob if
   the blob is a tree.

Changelog: fixed
EE: true