Speed up 13_comments seeder by up to 80%
This parallelizes the `13_comments` seeder, reducing the duration by
around 80% on macOS (~275 to ~55).

As identified previously, this fixture seeder is a major contributor to
the duration of all seeding. While we have added a `--fast` option to
reset a seeded GDK database in less than a minute, seeding from source
is still the default operation and also runs in CI.