Modify organization_push_rules sync triggers
This commit modifies PostgreSQL triggers and functions that were added
in https://gitlab.com/gitlab-org/gitlab/-/merge_requests/186908 to
synchronize the `organization_push_rules` table when there's a write
operation on the `push_rules` table. The initial trigger didn't copy
the `id` on the push_rules table to the `organization_push_rules` and
this led to out of sync data.

Changelog: added