Collect data on hidden SAML merge requests on dashboard
Collects data in Sentry on the number of visible merge requests against
the count to check if any are hidden. This could be caused by a SAML
issue which we want to collect the data on.

In the Sentry event we are also sending over whether the reauth notice
is visible just to make sure the banner is getting displayed correctly.