Strengthen sharding key validation in Snippet
Snippet seems to serve as a base clase to ProjectSnippet and
PersonalSnippet. They each have a basic presence validation on either
organization or project. They do not currently have a validation saying
that they should only ever have 1 of those though.

We should reflect what's in the database constraint for snippets:

`"check_82c1d40fab" CHECK (num_nonnulls(organization_id, project_id) = 1)`

That enforces an exclusive OR relationship - a snippet must belong to
either a project OR an organization, but not both and not neither.

This commit adds a validation to the Snippet class that can reflect
that.