Add shims for Gitlab::Saas in FOSS
Currently, the Gitlab::Saas module exists in FOSS, but has several
methods that only exist in EE. We should ensure that all methods are
available in both editions, with shims for FOSS-only. Otherwise we run
the risk of raising `NoMethodError` only in the FOSS codebase, which can
be hard to catch during regular development on EE and CE features.