Add an oauth_access_grant_archived_records table and retention policy
The purpose of this table is to store deleted rows from the operational
`oauth_access_grants` table. We will be archiving records from
`oauth_access_grants` and will store deleted records in the new table
as a soft-delete mechanism.

The archive table is meant to be temporary and we will remove it after
the retention policy has been enforced on `oauth_access_grants` and
is verified.

Changelog: added
Relates https://gitlab.com/gitlab-org/gitlab/-/issues/377995