Compare partial scans vs full scans
The current comparison logic only compares partial scan results to
partial scan results. When there is a full scan on the default branch,
it won't be compared against. This is good when we don't want to show
fixed vulnerabilities, however it has a problem in that if an existing
vulnerability is found  by a partial scan, it will show as "new" even if
it's not. To fix this, we compare partial scan results vs full scan
results, but hide the fixed findings.