refactor(spec): Only check for validated NOT NULL constraints
The sharding key spec helpers `has_null_check_constraint?` and
`has_multi_column_null_check_constraint?` previously checked for the
existence of a constraint without confirming if it was validated by the
database.

An unvalidated (`NOT VALID`) constraint does not guarantee that existing
rows conform to the constraint. This could lead to sharding key specs
passing even when the underlying data is not yet backfilled and
guaranteed to be present.

This change adds `AND convalidated = true` to the SQL queries in
these helpers to ensure we only recognize constraints that are fully
enforced, making our sharding key validation more accurate.

Changelog: other