Improve title copy for user status form emoji picker
The default button title for the emoji picker component is "Add
reaction" but that is not an appropriate or descriptive title for the
button's role when selecting an emoji in the user status form.

This MR adds an override prop for the emoji picker's button title, and
uses that prop to set the title to "Choose emoji" when used in the user
status form.

Changelog: changed