Implement name and sha component context interpolation
This is the second step of adding component context to the
CI interpolation. In the first step, we've prepared the codebase by
refactoring YAML context passing. In this step, we are adding
`component.name` and `component.sha` to the interpolation. In the next
step, we'll add `component.version`.

The documentation will be added in the final step.

This change is behind the feature flag
`ci_component_context_interpolation`.