Explicitly skip Ci::Ref status updates for dangling pipelines
This adds a fast path to `Ci::Ref#update_status_by!(pipeline)` for
dangling pipelines. The pipeline for `#last_finished_pipeline_id` is
already guaranteed not to be dangling, so it will never equal
`pipeline.id` when for a dangling `pipeline`, but that's not immediately
obvious from the code, and requires a query that is redundant in this
special case.

No tests were changed because we already have a test case for this
scenario.

See https://gitlab.com/gitlab-org/gitlab/-/issues/571094