fix: Fix the path to the gitlab-docs project
While working on https://gitlab.com/gitlab-org/gitlab/-/merge_requests/205960, we saw a failing
[docs-lint redirects CI job]. The job's failure message mentioned the `gitlab/gitlab-docs` project
which does not exist. I found gitlab-org/gitlab-docs but that is deprecated in favor of
gitlab-org/technical-writing/docs-gitlab-com. Linking to the right project in this failure message
will help team members who are unfamiliar with this CI job to find out how to fix it when it fails.

[docs-lint redirects CI job]: https://gitlab.com/gitlab-org/gitlab/-/jobs/11451202235