Adjust "Run pipeline" responsive behavior
Adjust the "Run pipeline" button's responsive behavior so that in the
`md` range the button is not full-width, and add appropriate spacing
above and to the left and right of the button. This MR also adjusts the
button variant to use the default variant, to match what is used within
the table header.

Changelog: changed