Introduce `ignoreRevs` argument to GraphQL `blame` query
Gitaly now supports filtering revisions out of the blames using
ignore-revisions-blob.

We will now allow the frontend to send a toggle
to enable or disable this filtering.

If the file doesn't exist or is corrupt a warning will be returned to
the client.

Related to https://gitlab.com/gitlab-org/gitlab/-/issues/519397

Changelog: added