Wrap column names in quotes
Similar to https://gitlab.com/gitlab-org/gitlab/-/merge_requests/149351,
there are some examples in the existing DB where column names are
reserved words, and are defined on tables with quotes (specifically, on
merge_request_diff_files there is a column -- "binary"). Changing the
column name might be a Good Idea At Some Point, but for now, wrapping
the column names here in quotes will allow for migrations required as
part of on-going partitioning work to proceed, without having to wait
for several releases while we work through a lengthy renaming process.