Implement MCP client that uses Gitlab MCP server
MCP client is implemented in Workhorse:

- It enables support for self managed customers
- It enables support for customers with IP restrictions
- It enables support for airgapped customers

Currently, only Gitlab server configuration with a single tool
support (`get_issue`) is specified; however, the implementation
designed in a way that allows specifying other MCP server
configurations as well.

Additionally, `x-gitlab-enabled-mcp-server-tools` header is
passed that specifies which tools are now expected to be called
from the GitLab MCP server. It'll allow us to gradually migrate
from the tools specified in DWS to the MCP server tools.

The implementation is behind `mcp_client` feature flag.