Enable request store for seeders in development
This enables the `Gitlab::SafeRequestStore` for seeders in development
to cache the application settings as intended.

Doing this saves about 7ms every time a seeder reads the application
settings (e.g. when using `can?`), which accounts for about 30% of all
CPU time within seeders, thus speeding up seeding by around 30%.

On an M3 Max, this reduces the time to seed from around 320s to 215s
(~32%).