Mock sleep in gitlab:usage_data task spec
This mocks a `sleep` in the spec for `lib/tasks/gitlab/usage_data.rake`.

The sleep in the `GitlabServicePingWorker` is intended for a production
environment where running the task immediately after being started could
cause negative performance implications. This explicitly doesn't happen
in a test environment.

This speeds up the spec file by 1 minute on average.