Remove sessions fixed height
We had a fixed height on sessions to allow
the header to remain in place, but we are
currently supportin 3 different layout and
we keep seeing regressions in each one.

To prevent this, we now let the DOM flow
naturally, which will simplify the UI quite
a lot and work in all env.