Add another exception string for a wiki example
I can't think of a better way to work around the problem that's
happening in 206751. Marcin needs another string to use as an
exception, and this rule can't be disabled. I'm unlikely to find a
fast fix for implementing an internal-links rule in markdownlint,
so a clunky exception might be in order here.