Query commits metadata table to check if commit exists
In `Projects::MergeRequests::ApplicationController#commit`, we
check if commit exists in the DB by given `sha`. We are currently
querying the `merge_request_diff_commits` table.

Since we'll be moving the `sha` columne eventually to the
`merge_request_commits_metadata` table, we need to be able to
query that table as well.

We need to support both tables since there'll be records that
won't have associated `merge_request_commits_metadata_id` until
we backfill all the records.

This is behind `merge_request_diff_commits_dedup` feature flag.