Fix wiki inaccessible when deleting and creating group
This MR expires group wiki repository cache after getting
Gitlab::Git::Repository::NoRepository in response to accessing the wiki
repository through Gitaly.

This fixes a bug, where a group is deleted and then recreated with the
same name, and the wiki repository is inaccessible until the cache
expires.

Changelog: fixed