Fix pipeline table top borders
This MR addresses two problems:

1. At large container width and higher, there is a gap between the
bottom of the tabs bar and the top of the table, leading to two lines
being visible. This is fixed by making the entire offending div
conditionally displayed instead of its contents.
2. Even when that problem is addressed, the table's top border and the
tab bar's bottom border are right against each other, forming a
double-thickness border. This is fixed by adding a top margin of -1px to
the pipeline table when in this breakpoint range, so that only a single
1px apparent border is visible.

Changelog: fixed