Add Workhorse health check listener
This commit adds support for a health check listener that handles
readiness probes, designed specifically for Kubernetes. It can
optionally talk to the Puma control app and determine whether the app
has booted.

To enable this, add this to the Workhorse `config.toml`:

```
[health_check_listener]
  network = "tcp"
  # Address to bind the health check server to
  addr = "localhost:8082"
  readiness_probes_url = "https://gdk.test:3443/-/readiness"
```

When configured, the readiness probe will:

1. Attempt to contact the Puma control server (if specified with
`puma_control_url`). If that shows no workers have booted, then
the readiness will be set to `false`.

2. If the control server is up, then the health check will attempt
to probe the upstream `/-/readiness` endpoint.

The health checker listener has `max_consecutive_failures`,
`min_successful_probes`, and `check_interval` that can be used to tweak
the thresholds of when the readiness endpoint is marked as ready or not
ready. Setting `max_consecutive_failures` and `min_successful_probes` to
1 will flagging ready or not ready immediately.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/324882