rules/rule/changes: Use changed paths
* Switch rule/changes to leverage Pipelines changed paths method.
This uses the FindChangedPaths gitaly RPC, which is not subject to
diff file limit constraints

* Allow push to report the paths changed by the push via the Gitaly
FindChangedPaths RPC. This differs from the expensive "modified_paths"
api that invokes the more expensive DiffStats RPC and is always
computed within the current push context.

* Introduce new method to the ExternalPullRequest model that reports all
changed paths within the external PR via the Gitaly FindChangedPaths
RPC. This differs from the modified_paths method in that it uses the
FindChangedPaths RPC rather than the DiffStats RPC, which is more
costly.

* Introduces a changed_paths interface to the Pipeline model, exposing
all changed paths within a given pipeline context. This is
accomplished by leveraging the Gitlay FindChangedPaths RPC via the
repository model. This differs from modified_paths_since in that it
does not required a base commit to determine modification from a
certain point and instead operates within the full pipeline context.
Additionally this uses the FindChangedPaths RPC not the more expensive
DiffStats RPC.

Changelog: changed