Add project_id columns and foreign keys to x509 tables
Contributes to:
* https://gitlab.com/gitlab-org/gitlab/-/issues/553481
* https://gitlab.com/gitlab-org/gitlab/-/issues/553480

This migration adds `project_id` columns to both `x509_issuers` and
`x509_certificates` tables along with their foreign key constraints
to the projects table and index.

These changes enable proper sharding configuration for the x509
tables as part of the protocells initiative, allowing them to be
classified as org-level tables with project_id as the sharding key.

Changelog: added