Create risk score calculation service
This MR introduces the new
Vulnerabilities::Findings::RiskScoreCalculationService.
This service takes in one or more vulnerability ids
and updates risk scores for the corresponding
vulnerability_occurrences records. This is done
through an upsert_all operation so records are
either created or updated. This service is called
in places where severity changes or a vulnerability
transitions to an active state.

EE: true
Changelog: added