Adds the gl_user_id field to the application_context
This change introduces a new feature flag that controls
whether we set the existing meta.user_id or the new
meta.gl_user_id field.

This change falls in line with our new defined standards for
logging field names. By publishing both, we will be increasing the
number of fields within our logs, however, this is intentional
as it will facilitate an easier migration of all non-standard
fields over to the new one.