Remove `tailwind_container_queries` feature flag
Removes the `tailwind_container_queries` feature flag in favor of the
`project_studio_enabled?` condition. This means that the container
queries migration is now tied to the overarching Project Studio
initiative.
As the migration is more or less complete, and as we're getting ready to
start rolling out the new UI, regrouping both flags gives us an easier
way to toggle between relevant states.

Changelog: other