Move data-testid in placeholder_reassigned_actions component
Previously, the data-testid was duplicated on both the GlAvatarLabeled
root element and the inner img tag.

After the change in https://gitlab.com/gitlab-org/gitlab-services/design.gitlab.com/-/merge_requests/5104
the data-testid appears only in the img tag,

This change moves the data-test-id to the parent element to fix the E2E
test.