Agentic VR badges
As part of the initiative to add an agentic flow for
resolving SAST vulnerabilities, we are adding new
badges and states to the vulnerability report.

- AI fix in progress badge
- AI fixed badge

This change is behind a new feature flag `agentic_sast_vr_ui`.

The AI fix in progress badge expects a yet to be implemented
graphql field `aiFixInProgress` which will be based on related
workflows. This now uses a client-side resolver to simulate
some of the vulnerabilities having a fix in progress.