Restrict Puma control status app commands
In Cloud Native GitLab, we want to enable the Puma HTTP control app as
part of the Kubernetes readiness check, but the control app also allows
a user to restart, stop, refork, and halt a Puma process. We want to
restrict the functionality to only allow retrieval of data in `/stats`.
The Puma maintainers appear to be amenable to
https://github.com/puma/puma/pull/3787, but we need to lock this down
ASAP for https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/2654.

Relates to https://gitlab.com/gitlab-com/gl-infra/production/-/issues/20469