Delete the `show_container_registry_tag_signatures` feature flag
This feature flag was introduced in 17.1, which is more than 4 milestones ago.

As part of our process we want to ensure [feature flags don't stay too long in the codebase](https://docs.gitlab.com/ee/development/feature_flags/#types-of-feature-flags).

Rollout issue: https://gitlab.com/gitlab-org/gitlab/-/issues/460226

<details><summary>Remaining mentions of the feature flag (click to expand)</summary>

```
CHANGELOG.md
13688:- [Enable show_container_registry_tag_signatures feature flag by default](gitlab-org/gitlab@3d600c9a74c75d04257f2ff4baa6d49d8b5b5b54) ([merge request](https://gitlab.com/gitlab-org/gitlab/-/merge_requests/154706)) **GitLab Enterprise Edition**

```

</details>

 **Currently the feature flag is `enabled` on production**

It is possible that this MR will still need some changes to remove references to the feature flag in the code.
At the moment the `gitlab-housekeeper` is not always capable of removing all references so you must check the diff and pipeline failures to confirm if there are any issues.
It is the responsibility of ~"group::container registry" to push those changes to this branch.

**Note:** If you do not want to remove this feature flag at this time, you can add an `intended_to_rollout_by_date` attribute in the feature flag YAML file to prevent automated removal.

## TODO for the reviewers before merging this MR
- [ ] See the status of the rollout by checking https://gitlab.com/gitlab-org/gitlab/-/issues/460226, https://gitlab.com/gitlab-com/gl-infra/feature-flag-log/-/issues/?search=show_container_registry_tag_signatures&sort=created_date&state=all&label_name%5B%5D=host%3A%3Agitlab.com
- [ ] Verify the feature flag status via chatops by running `/chatops run feature get show_container_registry_tag_signatures`.
- [ ] [Search for references to `ShowContainerRegistryTagSignatures` in frontend part of code](https://gitlab.com/search?project_id=278964&scope=blobs&search=ShowContainerRegistryTagSignatures&regex=false)
- [ ] [Search for references to `show_container_registry_tag_signatures` in code](https://gitlab.com/search?project_id=278964&scope=blobs&search=show_container_registry_tag_signatures&regex=false)
- [ ] Check if we need to remove any Gem or other related code by looking at the changes in https://gitlab.com/gitlab-org/gitlab/-/merge_requests/152135

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
in [CI](https://gitlab.com/gitlab-org/quality/engineering-productivity/team/-/jobs/11531420999) using the `Keeps::DeleteOldFeatureFlags` keep.

To provide feedback on your experience with `gitlab-housekeeper` please create an issue with the
label ~"GitLab Housekeeper" and consider pinging the author of this keep.

Changelog: removed