Fix missing gitaly_context forward in BranchPushService
In order to prevent "CI loops" with the job token push feature, we set a
`glBuildId` in the Gitaly context whenever a push originates from a CI
job, and forbid pipeline creation when this key is present. This was
already working for branches and tags, but we missed forwarding this
variable for merge request pipelines.

See https://gitlab.com/gitlab-org/gitlab/-/merge_requests/197418#note_2745993919

Changelog: fixed