Ensure skiplink is first focusable element
The new super top bar in paneled UI is earlier in the DOM than the super
sidebar, which means that its focusable elements come before the super
sidebar's focusable elements. This means the skiplink button in the
super sidebar cannot be used when paneled UI is enabled.

This MR adds a new skiplink button to the super topbar for paneled UI.

Changelog: fixed