Check granular permission for admin users
When admin users authenticate to access resources in
the admin area the granular permission is now checked
instead of the broad admin check.

Additionally, users with a custom admin role that grants
access to an admin area resource are redirected to enable
admin mode, if needed, instead of presented a 404.

Changelog: changed