Fix MergeRequestsFinder commit filtering for generated ref commit shas
Prior to this change, that commit sha could widen the scope of the merge
request finder if they only matched a generated ref commit sha. The fix
is to move the handling of the generated ref commit shas from the finder
and into the scope `MergeRequest.by_related_commit_sha`, for consistency
with existing commit filtering.

Changelog: fixed