Fix layout of AI panel
The AI Panel body currently uses `gl-h-full`, but it actually should not
be the full height of its parent, as it has a prior sibling. This causes
some unpleasant overlapping.

Instead, this MR changes the `<aside>` to be laid out as a flex column,
and adds `gl-grow` to the ai panel body so that it will fill all of the
available vertical space.

Changelog: fixed
EE: true