Update AWS SDK for Workhorse and disable AWS checksums by default
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/198371, we had
to downgrade the AWS SDK because AWS enabled default integrity
protections by default that are incompatible with third-party S3
providers, such as Linode.

Previously it was only possible to disable this behavior partially with
the `AWS_REQUEST_CHECKSUM_CALCULATION` and
`AWS_RESPONSE_CHECKSUM_VALIDATION` environment variables, but this did
not fix the S3 uploaders that used the SDK's `manager.Uploader`
implementation.

With https://github.com/aws/aws-sdk-go-v2/pull/3151 is it now possible
to disable this default behavior with `manager.Uploader`.

To ensure backwards compatibility with the many S3 providers that GitLab
supports, disable these default integrity protections for now. We should
look to re-enabling them when support for this is more common. For now,
admins can set the environment variables if they want to re-enable this
behavior.

Changelog: changed