Add local cache entry table and model to Maven virtual registries
Introduce a local cache entry concept to Maven virtual registries.
It is simply a link between an upstream (in local mode) and an
existing package file.

If the upstream, group or package file gets destroyed, the existing
records in the table are safe to be deleted by the database cascading
deletes.

Changelog: added
EE: true