Fix bug in UserJWT CEL policy for Secrets Manager
Prior to this change, we had an issue where the
UserJWT CEL policy for the Secrets Manager
was not correctly validating the subject claim,
leading to potential security risks. We were also lacking tests for
various JWT based authentication scenarios with Openbao. Adding the
tests helped uncover this authentication vulnerability.

This change fixes the issue by adding a check
to ensure that the for authentication via UserJWT,
we check for the subject claim as well. We added
comprehensive tests to cover various JWT authentication scenarios
with Openbao.

References
[Issue 1](https://gitlab.com/gitlab-org/gitlab/-/issues/575336)
[Issue 2](https://gitlab.com/gitlab-org/gitlab/-/issues/548595)

Changelog: changed
EE: true