Update OldestPerCommitFinder to query SHA from metadata table
This finder uses the method
`MergeRequestDiffCommit.oldest_merge_request_id_per_commit`
which queries the table `merge_request_diff_commits` by sha.

Since we'll be moving the `sha` column to the
`merge_request_commits_metadata` table, we need to be able to
query that table as well. To achieve this we use a similar
method in the `MergeRequests::CommitsMetadata` which queries
and group the result using the `sha` column in the metadata
table.

We need to support both tables since there'll be records that
won't have associated `merge_request_commits_metadata_id` until
we backfill all the records.

This is behind `merge_request_diff_commits_dedup` feature flag.