Change sleep seconds
Since we uped the TTL to 60 seconds
this will help ensure we do not return
conflict for some statuses when another
status is still being created. By spacing out
the timeframe between retries we give this a
chance to recover over 10 seconds instead of
2 seconds

Update the docs with a status code.