Update method `MergeRequest#all_commit_shas`
In `MergeRequest#all_commit_shas`, we query commits SHAs from
the `merge_request_diff_commits` table.

Since the column `merge_request_diff_commits.sha` will be moved
to the `merge_request_commits_metadata` table, we need to update
the query to look in this table as well.

We revert to the original behaviour when the feature flag
`merge_request_diff_commits_dedup` is disabled or the table
`merge_request_commits_metadata` has no records for the
given commits.