Downgrade redis-rb gem to v5.4.0
In https://github.com/redis/redis-rb/pull/1312, `redis` v5.4.1 was
updated to translate `RedisClient::ConnectionError` to
`Redis::ConnectionError` to avoid leaking the underlying
exception. However, ActionCable v7.1.5.2 relied on this behavior and
only reconnected if `RedisClient::ConnectionError` were raised.  As a
result, if the Redis server terminated an idle connection, the Puma
thread would raise an exception and cause the worker process to restart.

Rails has since been updated to handle this in
https://github.com/rails/rails/pull/55359. The fix has been committed to
the Rails `7-1-stable` branch but not yet released.

Relates to https://gitlab.com/gitlab-com/request-for-help/-/issues/3593

Changelog: fixed