Ignore organization_id on system_note_metadata
system_note_metadata was originally supposed to directly
inherit it's sharding key from the notes table. Afterwards,
we realized that we don't have system notes for personal snippets
and that was the only reason to have the organization_id column.

With this change we can reuse a simpler function that doesn't
account for organization_id presence on the note's table sharding key

Changelog: other