Recreate indexes that were inadvertently deleted
The following merge requests deleted unneeded indexes for GitLab.com:

- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/122368
- https://gitlab.com/gitlab-org/gitlab/-/merge_requests/157136

These were removed with an async process, but they were not supposed
to be deleted for self-managed instances in part for usage pings. This
happened because:

1. Async index removal is done through the `gitlab:db:reindex` rake task
(this is something we need fix as it was supposed to be a temporary
solution).

2. With https://gitlab.com/gitlab-org/gitlab/-/merge_requests/188548 we
enabled the `database_reindexing` flag by default.

3. The migration that scheduled the indexes for removal does not have a
guard to execute only for GitLab.com.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/567561