AI Catalog permissions and `duo_features_enabled`
Currently AI Catalog features are enabled if the project has opted-
in to Experimental/Beta AI features. This check will eventually be
removed when AI Catalog goes GA.

When the check is removed, currently it would make AI Catalog features
available to any project that is Premium+.

This change is to check that the project has `duo_features_enabled`,
which is a cascading setting that can be set at instance, group, or
project-level to signal if Duo features should run at that level.

Part of:
https://gitlab.com/gitlab-org/gitlab/-/issues/568176