Add secrets_manager_scope claim for SecrtsMgr Auth
We have 3 types of Auth mounts and JWTs to
authenticate with these auth mounts in Openbao - User, Project and Pipeline.
Openbao. Prior to this change, we did not have
an explicit way to differentiate between these types of JWTs in our codebase.

This change fixes the issue by adding a new bound
claim check `secrets_manager_scope` to all 3 JWT auth roles in
Openbao. We also add this claim to all 3 types of JWTs that we
generate in GitLab - UserJwt, ProjectJwt and PipelineJwt.

References:
[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/576441)

Changelog: changed
EE: true