Add tag and latest_pipeline_per_ref to pipelines
Related to https://gitlab.com/gitlab-org/gitlab/-/issues/347526+

The branches and tags pages show a status icon/link to the latest
pipeline for the ref. Currently this points to the latest pipeline for
the commit this ref points to which is incorrect.

Currently if you have a tag and branch with the same name e.g.
`refs/heads/v1.1.0` and `refs/tags/v1.1.0` we do not have a scope to
filter for tag pipelines, we do however have a `no_tag` scope.

The new `tag` scope will allow us to filter the pipelines for the
correct type.

The new `for_ref_and_sha_pairs` scope will accept a list of tuple arrays
of refs and shas, e.g. `[['master', '1ba497f'], ['dev', '9abd10f]]`,
which are then used to locate pipelines at a specific location within
each ref.

Currently we use the latest_pipeline_per_commit method by passing all of
the commits for the refs but this means we show the latest pipeline for
the commit not the ref itself.
The new `latest_pipeline_per_ref` method takes a list of refs and shas
e.g. `[['master', '1ba497f'], ['dev', '9abd10f]]`.
The method will use this to find pipelines for the refs at those
specific shas.

We build a join to optimize how postgres utilizes the available
indexes. I attempted an OR query and a tuple matching query and both
resulted in poor performance for any more than 20-30
refs sha pairs, reaching up to 800ms with 100 pairs. The current query
performs around 15-30ms for 100 pairs for tags and branches.

These queries will be used in a future change to
Ci::CommitStatusesFinder