Always return RateLimit-* headers
We use RackAttack for rate limiting purposes.
It returns RateLimit-* headers for throttled requests, but not for
unthrottled ones.

This adds a middleware to always return RateLimit-* headers, which
allows a client to anticipate when it'll run into a rate limiting
situation and act accordingly before this happens.

Related to https://gitlab.com/gitlab-com/gl-infra/production-engineering/-/work_items/25770