Add rake task for partition mangement
Given that data retention is a high organizational priority, the ability
to detach partitions before dropping them gives us a path to "soft
delete" so we can evaluate if we need to actually keep the data before
deleting it. Also given that we need to be able to detach and quickly
reattach a partition if the detaching causes an incident, it's better to
provide a tool to run this operation from the command like as opposed to
accomplishing this through migrations. This tool allows us to run
partition detachment operations from the command line.

Adds two new rake commands:

detach_partition[partition_name] - Verifies that:

1. Are allowed to detach the partition
2. That we have all of the information and setup needed to detach a
   partition

and then detaches the partition.

reattach_partition[partition_name]

Reattaches the partition to its parent table.