Mark 20250516124136 as obsolete
This migration marks the 20250516124136 BackfillWorkItemMilestoneData Advanced search migration as obsolete.

[Search for references to `backfill_work_item_milestone_data` in code](https://gitlab.com/search?project_id=278964&scope=blobs&search=backfill_work_item_milestone_data&regex=false))

At the moment, the `gitlab-housekeeper` is not always capable of removing all references so
you must check the diff and pipeline failures to confirm if there are any issues.
It is the responsibility of the assignee (picked from ~"group::global search") to push those changes to this branch.

[Read more](https://docs.gitlab.com/ee/development/search/advanced_search_migration_styleguide.html#cleaning-up-advanced-search-migrations)
about the process for marking Advanced search migrations as obsolete.

All Advanced search migrations must have had at least one
[required stop](https : // docs.gitlab.com / ee / development / database / required_stops.html)
to process the migration. Therefore we mark any Advanced search migrations added before the
last required stop as obsolete.

This change was generated by
[gitlab-housekeeper](https://gitlab.com/gitlab-org/gitlab/-/tree/master/gems/gitlab-housekeeper)
in [CI](https://gitlab.com/gitlab-org/quality/engineering-productivity/team/-/jobs/11762011156) using the `Keeps::MarkOldAdvancedSearchMigrationsAsObsolete` keep.

To provide feedback on your experience with `gitlab-housekeeper` please create an issue with the
label ~"GitLab Housekeeper" and consider pinging the author of this keep.

Changelog: other
EE: true