Delegate MRDC methods to commits metadata when present
When a `MergeRequestDiffCommit` record has an associated
`MergeRequest::CommitsMetadata` record, we should be getting
the value of some methods from that record instead.

This is needed because when we stop writing to the columns in
`merge_request_diff_commits` table, the values should be read
from the `merge_request_commits_metadata` table.

The following columns that are delegated:
- `commit_author`
- `committer`
- `authored_date`
- `committed_date`
- `sha`
- `message`

Intentionally skipping `trailers` as we're working on removing
it.

This is behind `merge_request_diff_commits_dedup` feature flag.