Avoid checking out DB connections when checking feature flags in specs
When feature flags are stubbed in specs, we do not need to do a database
existence check.

This prevents flaky specs when a connection is checked out in between
test teardown and setup of the next test. When a separate connection is
checked out, the transaction used by let_it_be / before_all is broken.