Reduce cached SQL queries in `/api/v4/internal/pages` endpoint
Previously `VirtualDomain#pages_access_control_trie` dominated the CPU
profile because the memoization was not working properly. It was being
memoized with the value of `self_and_descendants`, which is a different
object ID each time the method is called. Fix this by memoizing
`pages_access_control_trie` directly.

Changelog: performance