Fix Gitlab::Graphql::OffsetPaginatedRelation#preload
`Gitlab::Graphql::OffsetPaginatedRelation` is just a `SimpleDelegator`
wrapping a relation. But some methods, like `#preload`, return a new
object which loses the offset pagination type hint for GraphQL. To fix
it, we simply re-wrap the preloaded relation.

This removes the flag `fix_graphql_offset_pagination_preloads` that was
used to control the initial rollout on Gitlab.com.

Changelog: fixed