Do not filter by activity for the severity count panels
This makes sure that when clicking the View link on the vulnerability counts
per severity panel on the new dashboard the vulnerability report is not
filtered by activity ALL, but rather by the default value, activity
Still detected. The vulnerability report filters out No longer detected
vulnerabilities by default, and the vulnerability counts per severity
do that as well.