Add audit event streaming for secrets manager
We need to log audit events for read and update actions on secrets
in both project and group contexts for the users to track the activities
related to secrets management. Since the amount of read secret events is
high, we will not be persisting these events in the database but the
update events will be stored in the database.

This commit adds the audit event definitions, adds a new internal API
endpoint to receive the audit events from OpenBAO, and a model to
structure the audit log data.

Changelog: added
EE: true