Add Secrets Manager test cases for forked projects
Currently, we DO NOT have test coverage for
Secrets Manager access control in the context of forked projects.
This commit adds those test cases to ensure that
access control mechanisms are correctly enforced
when dealing with forked projects in GitLab's Secrets Manager.

We have two cases to cover here:
1. A pipeline running in a forked project trying to access secrets
   from the original project.
2. A pipeline running in the original project trying to access secrets
   from the forked project.

In this commit, we implement test cases for both Authentication
and Authorization scenarios for these cases.