Fix gfm reference for namespaces with the same path
When we create a system note reference when an issue gets mentioned in
an epic, we created a relative reference, like "mentioned in
gitlab#X".

However, when the group and the issue have the same namespace path name,
it was an ambigious reference because the reference parser is looking
first to resolve namespace paths within the current parent namespace.

In the example above, when there is a group and a project called
`gitlab`, we'd create the system note on the `gitlab` project with
`mentioned in gitlab#X`. When looking up a namespace with the
name `gitlab` in the parent group of the issue, we'd find the group and
would reference some issue that has the IID 123.

This is not correct, since we wanted to reference the group level
issue/work item.

By generating an absolute reference in the case above
`mentioned in /gitlab#X` we remove the ambiguity as the reference
parser can find the correct group work item.

Changelog: fixed
EE: true