Pass filter parameters to csv export request
Currently the Vulnerability Report's CSV
export doesn't take any of the applied
filters into account.

As a first step, this MR adds the filters
as parameters to the POST request.

It also lays the foundation to expand the
feature in the future. For example disabling
the export button if the selected filter set
does not produce any results and therefor
the export would also be empty.